+++ src/main/java/blackjack/dto/DeckDTO.java
@@ -1,6 +1,6 @@
 package blackjack.dto;
 
-import blackjack.Player;
+import blackjack.model.player.Player;
 import java.util.List;
 
 public class DeckDTO {
