+++ src/main/java/blackjack/model/player/Player.java
@@ -1,4 +1,7 @@
-package blackjack;
+package blackjack.model.player;
+
+import blackjack.model.trumpcard.Deck;
+import blackjack.model.trumpcard.TrumpCard;
 
 public abstract class Player implements PlayerInterface {
     private static final String ERROR_NULL = "[ERROR] 입력된 이름이 없습니다.";
