+++ src/main/java/blackjack/model/trumpcard/TrumpCard.java
@@ -1,4 +1,4 @@
-package blackjack;
+package blackjack.model.trumpcard;
 
 import java.util.Objects;
