+++ src/test/java/blackjack/model/trumpcard/TrumpNumberTest.java
@@ -1,7 +1,6 @@
-package blackjack;
+package blackjack.model.trumpcard;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
@@ -9,6 +8,6 @@ public class TrumpNumberTest {
     @ParameterizedTest(name = "{0}에 8을 더하면 {1}이다")
     @CsvSource(value = {"2:10", "7:15", "25:33"}, delimiterString = ":")
     void sumTo_value(int value, int expected) {
-        assertThat(TrumpNumber.EIGHT.sumTo(value)).isEqualTo(expected);
+        Assertions.assertThat(TrumpNumber.EIGHT.sumTo(value)).isEqualTo(expected);
     }
 }
